add prettier format and fix few mistakes
