+++ package.json
@@ -35,6 +35,7 @@
     "karma-coverage": "~2.1.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
+    "prettier": "2.6.2",
     "typescript": "~4.6.2"
   }
 }
