+++ src/app/shared/data-storage.service.ts
@@ -1,7 +1,8 @@
 import { Injectable } from '@angular/core';
 import {HttpClient} from "@angular/common/http";
 import {environment} from "../../environments/environment.prod";
-import {RootObject} from "../weather/weather.service";
+import {RootObject} from "../weather/root-object.model";
+import {Observable} from "rxjs";
 
 @Injectable({
   providedIn: 'root'
@@ -10,7 +11,7 @@ export class DataStorageService {
   constructor(private http: HttpClient) {
   }
 
-  getCityWeather(city: String) {
-    return this.http.get<RootObject>(environment.openWeatherSpecificCityUrl + city + '&APPID=' + environment.apiKey)
+  getCityWeather(city: String): Observable<RootObject> {
+    return this.http.get<RootObject>(`${environment.openWeatherSpecificCityUrl}${city}&APPID=${environment.apiKey}`)
   }
 }
