+++ src/environments/environment.prod.ts
@@ -2,5 +2,6 @@ export const environment = {
   production: true,
   signInUrl: "https://identitytoolkit.googleapis.com/v1/accounts:signInWithPassword?key=AIzaSyCfgBbiH9i6-3Lg22h47RNhyWZegGlyZjk",
   openWeatherSpecificCityUrl: "https://api.openweathermap.org/data/2.5/weather?q=",
-  apiKey: "1eafa79098796a04151a38fb2c7f2499"
+  apiKey: "1eafa79098796a04151a38fb2c7f2499",
+  openWeatherImageURL: "http://openweathermap.org"
 };
